refactor: Use Rich Table for timing report
Rather than generating the timing report table manually, use
`rich.table.Table` to handle the magic automatically.  Also abandon the
Markdown format for the table, as Rich's default styling looks nicer,
and if a user needs to copy it into an issue, they can always place it
in a code block.