feat!: Allow stages to be run multiple times
Convert the `durations` attribute from a `dict` to a `list` to allow for
the possibility that one or more stages might be run multiple times
(e.g., if something failed).  Introduce a `StageDuration` class that
inherits from `NamedTuple`, such that the entries added to the list must
always be a `(str, timedelta)` `tuple` that will from then on be
immutable and allow easy member access via dot notation.