refactor!: Stages always return None
In preparation for adding stage retry functionality, make it such that a
function decorated with `DriverScript.stage` must return `None`.  This
is reasonable because a conceptual stage of a script *does* something,
but doesn't *return* something.

Note that this is a breaking change for `DriverScript`, as any
subclasses relying on returning values from functions decorated with
`stage` will no longer work.  A follow-up commit will fix this problem
for `GMSSystemTest`.