refactor: Get phase method rather than running it
In preparation for adding automatic retry functionality to the `stage`
decorator, rewrite the `run_phase` method such that instead of running
the method corresponding to the phase, it simply returns the `Callable`,
which is then run outside the function.  This generalization will make
it such that we can pass a phase method to the `Retrying` object when
we're ready for that.