feat!:  Add stage retry functionality
Add the ability to automatically retry a stage that raises a
`tenacity.TryAgain` exception:

* Provide default implementations for methods to prepare to retry a
  stage and appropriately handle when retrying ultimately fails, while
  allowing subclass developers to override them or customize them on a
  stage-by-stage basis.
* Provide documentation to instruct subclass developers on how to make
  use of the retry functionality.
* Wrap the **Begin-Stage Actions**, **Stage Body**, and **End-Stage
  Actions** in the retrying loop.

Note that this is marked as a breaking change, as it's unclear what will
happen to any subclasses that implemented their own stage retry
functionality outside of this.  The safest thing will be for subclass
developers to update their subclasses to use this capability now
provided by the base class.