fix!: Handle --dry-run correctly
When a `DriverScript` subclass is run in dry-run mode, instead of
executing commands in the shell, print a message indicating what would
have been run.  Prior to this, `DriverScript` provided the flag, but
relied on subclass developers to ensure commands weren't run.  This was
an oversight.

Note that this is a breaking change, as behavior of the `run()` method
has fundamentally changed for dry-run mode.