feat: Add raise_parser_error
Add a method to the `DriverScript` base class to raise a parser error
when a user has done something wrong in specifying the command line
arguments.  This differs from `argparse.ArgumentParser.error()` in that
it's a little more user friendly, printing the help text first, and then
the particular error in yellow.