feat: Add flag for printing commands
Add a flag when instantiating a `DriverScript` to govern whether
commands should be printed immediately before running them.  Default it
to `True` for backwards compatibility.  Also provide a keyword argument
to the `run()` method that, if specified, will override the class flag.
In this way users could, e.g., instantiate a `DriverScript`, telling it
to print all commands, and then specify that certain commands should not
be printed.  The opposite of this is also possible (e.g., don't print
any commands, except this one and that one).