fix: Retry group help text
When I created the retry group in the argument parser, I didn't realize
that I was using `add_argument_group` in a way in which it wasn't
intended to be used.  This commit brings things into conformity with
how `argparse` is intended to be used, specifying a one-word group name,
followed by a more detailed group description.