feat: Add script name/stem attributes
Capture the name and stem of the file being executed as attributes of
`DriverScript` such that subclass developers can easily refer to them
rather than determining them on the fly.  This is useful for things like
the script execution summary.

Also add a unit test for `raise_parser_error()`, which was accidentally
omitted in a prior commit.