feat!: Create RetryStage exception
Rather than relying on stages raising a `tenacity.TryAgain` exception to
trigger retrying a stage, add a `RetryStage` exception to the package.
This makes things more explicit for subclass developers, and makes it
such that something outside the subclass developer's control that
happens to raise a `TryAgain` doesn't trigger retying the stage, unless
of course the subclass developer catches it and then raises a
`RetryStage` in its place.

Note that this is a breaking change, because any subclasses relying on
raising `TryAgain` to trigger a stage retry will need to be updated to
raise the new exception.