refactor!: Switch stages to standard set
Initially the `stages` attribute of a `DriverScript` class was
implemented as an ordered set (which in Python is a value-less `dict`
converted to a `list`).  The thinking there was that a `DriverScript`
subclass should define a series of stages in order and they would be
executed in that order.  However, this failed to consider the following
use case:  Class `A` inherits from `DriverScript` and defines a few
stages, then class `B` inherits from `A` and defines few more, but the
intended sequence of stages isn't all from `A` followed by all from `B`;
instead, they're meant to be interspersed.  Since this is a valid use
case, it makes sense to switch `stages` over to a standard `set`.

Another motivation for the `stages` being an ordered set was considering
the possibility of wanting to return to a prior stage if a certain stage
failed.  Consider an example `DriverScript` subclass with stages to
configure, build, and test a code.  If the test stage fails, perhaps the
failure was seemingly random, and it'd be worthwhile to automatically
rewind and retry the build and test stages again before notifying anyone
of the failure.  Or perhaps a certain regex match in the test output
indicates you should rewind further and reconfigure.  However, after
further reflection, allowing such flexibility would open up a can of
worms that probably shouldn't be opened.  `DriverScript` already allows
a subclass developer to automatically retry a stage.  If something's
happening that would cause you to back up further than the current stage
and try things again, that probably means there's flakiness either in
your code base or its surrounding infrastructure that a human should
look into, so `DriverScript` probably shouldn't allow you to hide that.

Note that this might be a breaking change if any subclasses were relying
on the ordering of the `stages`.