refactor!: Stages no longer registered by default
Since its inception, `DriverScript`'s `stage` decorator has attempted
to capture all the `stage`s defined by `DriverScript` subclasses in the
`stages` class variable automatically.  The desire was to keep things as
easy as possible for subclass developers, in that all they'd need to do
was inherit from `DriverScript`, define some methods decorated with
`stage`, and then `DriverScript` would automatically handle things like
stage-specific arguments (retries, etc.) and methods (pre/post,
begin/end, etc.).  Unfortunately this feature has been a thorn in our
side for development, due to, e.g., `pytest` multiple-loading modules
(and therefore redefining stages), `DriverScript` subclass hierarchies,
etc.

In order to pave the way for the flexibility for one `DriverScript`
subclass to use stages defined within a separate `DriverScript`
subclass, this commit makes it such that stages are no longer
automatically registered.  Instead subclass developers must pass the set
of stages to register into the constructor when instantiating an object.
This is a minor inconvenience, but it removes a good deal of hacky code
and makes it such that `DriverScript` subclasses are decoupled from one
another, so we believe the change is well worth it in the long run.

Note that this is a breaking change, as all subclasses must be update
their constructor definitions and any instantiations.