chore: Ignore unchecked shell input
When we built the precursor to `staged-script`, we didn't understand the
security implications of having the user pass commands to the underlying
shell as a string rather than a list of strings.  It just seemed like a
better interface, to make it easy for the user writing their Python
scripts to simply wrap their bash commands.  However, this opens up a
danger for bad actors to cause problems via command injection.  We can
remove the vulnerability by switching to only allowing a list of strings
as input to `run()`, but that breaking change will need to wait till
another day when I have more time available.