refactor!: Retry argument attributes
Mypy doesn't like that the retry argument attributes are defined
dynamically when a class is instantiated, so it'll complain that they
don't exist when a subclass developer is trying to modify them when
overriding/extending the parser.  If instead of having separate
attributes per registered stage, we just have a `dict`` where the keys
are the stage names, mypy is happy.

Note that this is a breaking change, as it changes how subclass
developers customize the parser.