docs: Include output from first example in README
To better advertise what the package provides, without requiring the
reader to jump to (or further into) the documentation, include the
output from running the first example script with both stages defined in
the `README.md` file, which will be displayed on GitHub and PyPI, and
the `index.rst` file, which will be displayed on ReadTheDocs.

Note that this incurs a minimal maintenance burden, in that we'll need
to keep this output up-to-date with any changes to the package that
affect the output.